reuse code between cycloneDX json & xml encoders
